Move compiler to its own class and add interface
