+++ modules/common/src/main/kotlin/org/cafejojo/schaapi/common/Project.kt
@@ -1,0 +1,13 @@
+package org.cafejojo.schaapi.common
+
+import java.io.File
+
+/**
+ * A project.
+ */
+interface Project {
+    /**
+     * The directory containing the project.
+     */
+    val projectDir: File
+}
