+++ modules/common/src/main/kotlin/org/cafejojo/schaapi/common/ProjectCompiler.kt
@@ -1,0 +1,14 @@
+package org.cafejojo.schaapi.common
+
+/**
+ * Represents a project compiler.
+ */
+interface ProjectCompiler {
+    /**
+     * Compiles a project.
+     *
+     * @param project an uncompiled project
+     * @return a compiled project
+     */
+    fun compile(project: Project): Project
+}
