+++ modules/pipeline/java-maven-project-compiler/build.gradle
@@ -1,5 +1,6 @@
 dependencies {
     compile project(":common")
+    compile project(":java-maven-project")
 
     compile group: "org.apache.maven.shared", name: "maven-invoker", version: mavenInvokerVersion
     compile group: "org.zeroturnaround", name: "zt-zip", version: ztZipVersion
