+++ modules/pipeline/java-maven-project-compiler/src/main/kotlin/org/cafejojo/schaapi/projectcompiler/javamaven/ProjectCompiler.kt
@@ -1,0 +1,66 @@
+package org.cafejojo.schaapi.projectcompiler.javamaven
+
+import org.apache.maven.shared.invoker.DefaultInvocationRequest
+import org.apache.maven.shared.invoker.DefaultInvoker
+import org.cafejojo.schaapi.common.Project
+import org.cafejojo.schaapi.common.ProjectCompiler
+import org.cafejojo.schaapi.project.javamaven.JavaMavenProject
+import java.io.File
+
+/**
+ * Compiles a Java project using Maven.
+ */
+class ProjectCompiler : ProjectCompiler {
+    /**
+     * Compiles a project.
+     *
+     * @param project an uncompiled project
+     * @return a compiled project
+     */
+    override fun compile(project: Project): Project {
+        if (project !is JavaMavenProject) throw IllegalArgumentException("Project must be JavaMavenProject.")
+
+        runMaven(project)
+
+        project.classes = project.classDir.walk().filter { it.isFile && it.extension == "class" }.toList()
+        project.classNames = project.classes.map {
+            it.relativeTo(project.classDir).toString().dropLast(".class".length).replace(File.separatorChar, '.')
+        }
+        project.dependencies = project.dependencyDir.listFiles().orEmpty().toList()
+        project.classpath =
+            if (project.dependencies.isEmpty()) {
+                project.classDir.absolutePath
+            } else {
+                project.classDir.absolutePath + File.pathSeparator +
+                    project.dependencies.joinToString(File.pathSeparator) { it.absolutePath }
+            }
+
+        return project
+    }
+
+    private fun runMaven(project: JavaMavenProject) {
+        val request = DefaultInvocationRequest().apply {
+            baseDirectory = project.projectDir
+            goals = listOf("clean", "install", "dependency:copy-dependencies")
+            isBatchMode = true
+            javaHome = File(System.getProperty("java.home"))
+            pomFile = pomFile
+        }
+
+        val invoker = DefaultInvoker().apply {
+            setOutputHandler(null)
+            mavenHome = project.mavenDir
+            workingDirectory = project.projectDir
+        }
+
+        val result = invoker.execute(request)
+        if (result.exitCode != 0) {
+            throw ProjectCompilationException("`maven install` executed unsuccessfully: " + result.executionException)
+        }
+    }
+}
+
+/**
+ * Indicates that the compilation of a project was unsuccessful.
+ */
+class ProjectCompilationException(message: String? = null) : Exception(message)
