+++ modules/pipeline/java-maven-project-compiler/src/test/kotlin/org/cafejojo/schaapi/projectcompiler/javamaven/ProjectCompilerTest.kt
@@ -1,6 +1,7 @@
 package org.cafejojo.schaapi.projectcompiler.javamaven
 
 import org.assertj.core.api.Assertions.assertThat
+import org.cafejojo.schaapi.project.javamaven.JavaMavenProject
 import org.jetbrains.spek.api.Spek
 import org.jetbrains.spek.api.dsl.describe
 import org.jetbrains.spek.api.dsl.it
@@ -9,7 +10,7 @@ import java.io.File
 import java.io.FileNotFoundException
 import java.nio.file.Files
 
-internal class JavaMavenProjectTest : Spek({
+internal class ProjectCompilerTest : Spek({
     lateinit var mavenHome: File
     lateinit var target: File
 
@@ -49,7 +50,7 @@ internal class JavaMavenProjectTest : Spek({
             setUpTestFiles("/ProjectCompiler/no-sources", target)
 
             val project = JavaMavenProject(target, mavenHome)
-            project.compile()
+            ProjectCompiler().compile(project)
 
             assertThat(project.projectDir).isEqualTo(target)
             assertThat(project.classes).isEmpty()
@@ -64,7 +65,7 @@ internal class JavaMavenProjectTest : Spek({
             setUpTestFiles("/ProjectCompiler/simple", target)
 
             val project = JavaMavenProject(target, mavenHome)
-            project.compile()
+            ProjectCompiler().compile(project)
 
             assertThat(project.projectDir).isEqualTo(target)
             assertThat(project.classes).containsExactlyInAnyOrder(
@@ -83,7 +84,7 @@ internal class JavaMavenProjectTest : Spek({
             setUpTestFiles("/ProjectCompiler/dependencies", target)
 
             val project = JavaMavenProject(target, mavenHome)
-            project.compile()
+            ProjectCompiler().compile(project)
 
             assertThat(project.projectDir).isEqualTo(target)
             assertThat(project.classes).containsExactlyInAnyOrder(
@@ -104,7 +105,7 @@ internal class JavaMavenProjectTest : Spek({
 })
 
 private fun setUpTestFiles(resourceString: String, target: File) {
-    val projectURI = JavaMavenProjectTest::class.java.getResource(resourceString)
+    val projectURI = ProjectCompilerTest::class.java.getResource(resourceString)
         ?: throw FileNotFoundException("Could not find test resources at $resourceString.")
 
     val projectFiles = File(projectURI.toURI())
