+++ settings.gradle
@@ -13,6 +13,7 @@ rootProject.name = "schaapi"
 
 includeModule "", "common"
 includeModule "models", "jimple-library-usage-graph"
+includeModule "models", "java-maven-project"
 includeModule "pipeline", "java-maven-project-compiler"
 includeModule "pipeline", "jimple-evosuite-test-generator"
 includeModule "pipeline", "jimple-library-usage-graph-generator"
